test(activemodel): merge misc.test.ts Attributes into attributes.test.ts; delete misc.test.ts
misc.test.ts is now empty except for the Attributes describe — merge it
into attributes.test.ts and delete the file entirely.
